+++ tests/map_lookup_wrapped_call_key/main.rs
@@ -4,7 +4,7 @@ use std::thread;
 
 pub fn key() -> u32 {
 
-    return 1 as u32;
+    return 1;
 }
 
 fn main() {
