+++ tests/package_global_channel_make/main.rs
@@ -176,7 +176,7 @@ fn __go_init_globals() {
 
 pub fn size() -> i32 {
 
-    return 2 as i32;
+    return 2;
 }
 
 fn main() {
